Separate file generation function from main.rs
